requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
PyJWT==2.8.0
pytz==2023.3
serverless-wsgi==3.0.2
six==1.16.0
Werkzeug==2.3.7
